pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "musk"
version = "0.1.0"
description = "A Vision-Language Foundation Model for Precision Oncology."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: CC BY-NC 4.0",
]
dependencies = [
"torch==2.0.1",
"torchvision==0.15.2",
"timm==0.4.12",
"Pillow==10.3.0",
"blobfile",
"mypy",
"numpy",
"openpyxl",
"pandas",
"h5py==3.11.0",
"threadpoolctl==3.5.0",
"protobuf==3.20.1",
"pytest==6.2.5",
"requests",
"einops",
"scipy",
"ftfy==6.1.3",
"opencv-python==4.9.0.80",
"sentencepiece==0.2.0",
"torchmetrics==0.7.3",
"transformers==4.38.1",
"deepspeed==0.4.0",
"pytorch-lightning==2.2.1",
"nltk",
"rouge",
"scikit-image",
"accelerate",
"fairscale",
"ruamel.yaml",
"wandb",
]

[tool.setuptools.packages.find]
exclude = ["assets*", "benchmark*", "docs", "dist*", "scripts*", "tests*", "checkpoints*"]

[tool.wheel]
exclude = ["assets*", "benchmark*", "docs", "dist*", "scripts*", "tests*", "checkpoints*"]